+++ src/app/screens/Play/hooks/usePlayScreenAnimation.ts
@@ -106,6 +106,9 @@ export function usePlayScreenAnimation(args: {
         pieceCount >= HIGH_PIECE_COUNT_THRESHOLD && !isDragging && !inCompletionFlourish;
 
       if (throttleIdle && now - lastFrameTimeRef.current < IDLE_MIN_INTERVAL_MS) {
+        lastCompleteRef.current = st.isComplete;
+        lastPieceCountRef.current = st.placedCount;
+        setState(st);
         rafRef.current = requestAnimationFrame(tick);
         return;
       }
@@ -260,15 +263,9 @@ export function usePlayScreenAnimation(args: {
         performance.measure("render-frame", "render-frame-start", "render-frame-end");
       }
 
-      const placedCount = st.pieces.filter((p) => p.isPlaced).length;
-      if (
-        st.isComplete !== lastCompleteRef.current ||
-        placedCount !== lastPieceCountRef.current
-      ) {
-        lastCompleteRef.current = st.isComplete;
-        lastPieceCountRef.current = placedCount;
-        setState(st);
-      }
+      lastCompleteRef.current = st.isComplete;
+      lastPieceCountRef.current = st.placedCount;
+      setState(st);
 
       rafRef.current = requestAnimationFrame(tick);
     };
